requirements.txt
Django==5.0.2
django-role-permissions==3.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
pillow==10.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.7.0